Rename egocentric hand tracking module
