+++ docs/reference/optional-dependencies.md
@@ -10,7 +10,7 @@ Install extras based on the data and operations you use.
 | Extra | Enables |
 | --- | --- |
 | `hf` | Hugging Face datasets, Hub APIs, and HF filesystem helpers. |
-| `egocentric` | Egocentric hand tracking. |
+| `hand_tracking` | Hand tracking with ego-vision. |
 | `hdf5` | HDF5 reader support. |
 | `zarr` | Zarr reader and writer support. |
 | `mcap` | MCAP robotics log reader support, including ROS2, protobuf, and H.264 video decoding. |
@@ -26,5 +26,5 @@ pip install macrodata-refiner[hf,video]
 pip install macrodata-refiner[hf]
 pip install macrodata-refiner[hdf5,zarr]
 pip install macrodata-refiner[mcap]
-pip install macrodata-refiner[egocentric]
+pip install macrodata-refiner[hand_tracking]
 ```
