+++ pyproject.toml
@@ -39,7 +39,7 @@ hf = [
     "huggingface-hub>=1.4.1",
     "hf>=1.7.1",
 ]
-egocentric = [
+hand_tracking = [
     "macrodata-refiner[hf]",
     "macrodata-refiner[video]",
     "ego-vision[models]>=0.1.15",
